python-analysis/requirements.txt
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
requests>=2.32.0
python-dotenv>=1.0.0
scipy>=1.13.0
tabulate>=0.9.0
